Correctly assign 1-bit fields
Fixes issue about assigning the value 1 to a 1-bit wide bit-field in a struct, which is signed by default. In C, a signed 1-bit field can only represent 0 or -1, so assigning 1 gets truncated to -1.

Why is this happening?
Bit-fields in C are tricky:
signed int x:1; can only be 0 or -1
unsigned int x:1; can be 0 or 1